pyproject.toml
[project]
name = "macroterm"
version = "0.1.0"
description = "A TUI for economic data and releases"
requires-python = ">=3.11"
dependencies = [
    "textual>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
macroterm = "macroterm.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macroterm"]
